+++ components/Button/index.js
@@ -2,64 +2,7 @@ import PropTypes from 'prop-types';
 import React, {useEffect} from 'react';
 import styled, {withTheme} from 'styled-components';
 
-const parseFontColor = props => {
-  const {fontColor, buttonVariant, disabled, theme} = props;
-  if (disabled && buttonVariant !== 'filled') return theme.colors.neutral8;
-  if (fontColor) return theme.parseColorValue(props, 'fontColor');
-  if (buttonVariant === 'outlined' || buttonVariant === 'ghost') return theme.parseColorValue(props, 'buttonColor');
-  return theme.colors.white;
-};
-const parseFontFamily = props => {
-  const {inheritFontStyle, theme} = props;
-  if (inheritFontStyle) return 'inherit';
-  return `${theme.fonts.primary}`;
-};
-
-//Get button size (height)
-const getSize = props => {
-  const factor = props.theme.factors.margin;
-  const size = props.buttonSize;
-  if (isNaN(size)) return `height: ${size}`;
-  if (size < 4) return `height: ${factor * 4}px`;
-  return `height: ${factor * props.buttonSize}px`;
-};
-
-//Get button width variations from props
-const getWidth = props => {
-  const factor = props.theme.factors.margin;
-  const size = props.buttonWidth;
-  if (typeof size === 'string') return `width: ${size}`;
-  if (size) return `width: ${factor * size}px`;
-  if (props.fullWidth) return 'width: 100%;';
-  return 'width: max-content;';
-};
-//Get button variations from props (return style matching the variation)
-const getVariations = props => {
-  switch(props.buttonVariant) {
-    case 'outlined':
-      return `
-        background-color: transparent;
-        border-width: 1px;
-        border-style: solid;
-        border-color: ${props.theme.parseColorValue(props, 'buttonColor')};
-        &:disabled {
-          border-color: ${props.theme.colors.neutral8};
-        }
-      `;
-    case 'ghost':
-      return `
-        background-color: transparent;
-      `;
-    case 'filled':
-    default: 
-      return `
-        background-color: ${props.theme.parseColorValue(props, 'buttonColor')};
-        &:disabled {
-          background-color: ${props.theme.colors.neutral8};
-        }
-      `;
-  }
-};
+import {parseProps} from './propsParser';
 
 const StyledButton = styled.button`
   display: flex;
@@ -69,19 +12,6 @@ const StyledButton = styled.button`
   text-transform: uppercase;
   border: unset;
   cursor: pointer;
-  svg {
-    fill: ${props => parseFontColor(props)};
-    width: 24px;
-    height: 24px;
-  }
-  span {
-    margin-left: 8px;
-    margin-right: 8px;
-    color: ${props => parseFontColor(props)};
-    font-size: 14px;
-    font-weight: 400;
-    font-family: ${props => parseFontFamily(props)};
-  }
   &:disabled {
     cursor: unset;
     &:hover {
@@ -97,11 +27,12 @@ const StyledButton = styled.button`
     from {opacity: 100%;}
     to {opacity: 80%;}
   }
-  ${props => props.theme.parsePadding(props.theme, props)};
-  ${props => props.theme.parseRadius(props, 'borderRadius')};
-  ${props => getVariations(props)};
-  ${props => getSize(props)};
-  ${props => getWidth(props)};
+  ${props => parseProps('', props)};
+  ${props => props.xs && parseProps('xs', props)};
+  ${props => props.sm && parseProps('sm', props)};
+  ${props => props.md && parseProps('md', props)};
+  ${props => props.lg && parseProps('lg', props)};
+  ${props => props.xl && parseProps('xl', props)};
 `;
 
 const StyledAria = styled.a`
@@ -114,7 +45,8 @@ const Button = ({
   color,
   disabled,
   fontColor,
-  buttonWidth,
+  fontFamily,
+  width,
   fullWidth,
   leftIcon,
   onClick,
@@ -125,9 +57,15 @@ const Button = ({
   variant,
   loading,
   enterKey,
+  removeText,
   px,
   on,
-  path
+  path,
+  xs,
+  sm,
+  md,
+  lg,
+  xl,
 }) => {
 
   // Trigger to Handle enter keydown for forms
@@ -139,21 +77,70 @@ const Button = ({
     return () => enterKey && window.removeEventListener('keydown', handleKeyPress);
   });
 
+  //All props with $ are forwarded ONLY to components (Not to HTML like onClick)
+  const getXsProps = () => xs && ({
+    $color: xs.color ? xs.color : color,
+    $variant: xs.variant ? xs.variant : variant,
+    $radius: xs.radius ? xs.radius : radius,
+    $size: xs.size ? xs.size : size,
+    $width: xs.width ? xs.width : width,
+    removeText: xs.removeText ? xs.removeText : removeText,
+  });
+  const getSmProps = () => sm && ({
+    $color: sm.color ? sm.color : color,
+    $variant: sm.variant ? sm.variant : variant,
+    $radius: sm.radius ? sm.radius : radius,
+    $size: sm.size ? sm.size : size,
+    $width: sm.width ? sm.width : width,
+    removeText: sm.removeText ? sm.removeText : removeText,
+  });
+  const getMdProps = () => md && ({
+    $color: md.color ? md.color : color,
+    $variant: md.variant ? md.variant : variant,
+    $radius: md.radius ? md.radius : radius,
+    $size: md.size ? md.size : size,
+    $width: md.width ? md.width : width,
+    removeText: md.removeText ? md.removeText : removeText,
+  });
+  const getLgProps = () => lg && ({
+    $color: lg.color ? lg.color : color,
+    $variant: lg.variant ? lg.variant : variant,
+    $radius: lg.radius ? lg.radius : radius,
+    $size: lg.size ? lg.size : size,
+    $width: lg.width ? lg.width : width,
+    removeText: lg.removeText ? lg.removeText : removeText,
+  });
+  const getXlProps = () => xl && ({
+    $color: xl.color ? xl.color : color,
+    $variant: xl.variant ? xl.variant : variant,
+    $radius: xl.radius ? xl.radius : radius,
+    $size: xl.size ? xl.size : size,
+    $width: xl.width ? xl.width : width,
+    removeText: xl.removeText ? xl.removeText : removeText,
+  });
+
   const renderRoot = () => {
     return (
       <StyledButton
-        onClick={onClick}
+        on={on}
         px={px}
-        disabled={disabled}
-        buttonColor={color}      
-        fontColor={fontColor}
-        borderRadius={radius}
-        buttonVariant={variant}
-        buttonSize={size}
-        buttonWidth={buttonWidth}
         fullWidth={fullWidth}
         style={style}
-        on={on}
+        onClick={onClick}
+        disabled={disabled}
+        fontColor={fontColor}
+        removeText={removeText}
+        $color={color}   
+        $variant={variant}
+        $radius={radius}
+        $size={size}
+        $width={width}
+        $fontFamily={fontFamily}
+        xs={getXsProps()}
+        sm={getSmProps()}
+        md={getMdProps()}
+        lg={getLgProps()}
+        xl={getXlProps()}
       >
         {loading && 'Carregando...'}
         {leftIcon && leftIcon}
@@ -190,7 +177,7 @@ Button.propTypes = {
   /**
    * Permite alterar a largura do componente (disponível a família de tamanhos em Number e String)
    */
-  buttonWidth: PropTypes.bool,
+  width: PropTypes.oneOfType([PropTypes.bool, PropTypes.string]),
   /**
    * Permite a passagem de um componente SVG para ícone no lado esquerdo
    */
@@ -202,7 +189,7 @@ Button.propTypes = {
   /**
    * Permite a escolha do token para border-radius
    */
-  radius: PropTypes.oneOf([false, 'default', 'alternative']),
+  radius: PropTypes.oneOf(['unset', 'default', 'alternative']),
   /**
    * Permite a passagem de um ícone à direita
    */
@@ -231,6 +218,14 @@ Button.propTypes = {
    * Permite alterar o espacamento interno no botão
    */
   px: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+  /** 
+   * Permite remover o texto do botão
+   */
+  removeText: PropTypes.bool,
+  /**
+   * Permite escolha de uma das propriedades do theme para font-family
+   */
+  fontFamily: PropTypes.string,
   /**
    * AMP :: Props equivalente a onClick, executa a ação do clique no AMP
    */
@@ -239,6 +234,46 @@ Button.propTypes = {
    * AMP :: Props equivalente a onClick, executa a ação de clique para links externos
    */
   path: PropTypes.string,
+  xs: PropTypes.shape({
+    color: PropTypes.string,
+    variant: PropTypes.oneOf(['filled', 'outlined', 'ghost']),
+    radius: PropTypes.oneOf(['unset', 'default', 'alternative']),
+    size: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    removeText: PropTypes.bool
+  }),
+  sm: PropTypes.shape({
+    color: PropTypes.string,
+    variant: PropTypes.oneOf(['filled', 'outlined', 'ghost']),
+    radius: PropTypes.oneOf(['unset', 'default', 'alternative']),
+    size: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    removeText: PropTypes.bool
+  }),
+  md: PropTypes.shape({
+    color: PropTypes.string,
+    variant: PropTypes.oneOf(['filled', 'outlined', 'ghost']),
+    radius: PropTypes.oneOf(['unset', 'default', 'alternative']),
+    size: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    removeText: PropTypes.bool
+  }),
+  lg: PropTypes.shape({
+    color: PropTypes.string,
+    variant: PropTypes.oneOf(['filled', 'outlined', 'ghost']),
+    radius: PropTypes.oneOf(['unset', 'default', 'alternative']),
+    size: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    removeText: PropTypes.bool
+  }),
+  xl: PropTypes.shape({
+    color: PropTypes.string,
+    variant: PropTypes.oneOf(['filled', 'outlined', 'ghost']),
+    radius: PropTypes.oneOf(['unset', 'default', 'alternative']),
+    size: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    removeText: PropTypes.bool
+  }),
 };
 
 Button.defaultProps = {
@@ -248,7 +283,8 @@ Button.defaultProps = {
   color: 'primary1',
   radius: 'default',
   size: 4,
-  loading: false
+  loading: false,
+  fontFamily: 'primary'
 };
 
 export default withTheme(Button);